⚙ refactor: refactor the code structure of the project
